[api] Refactor the Api from the WebServerApi module
